Additional float to int type conversions (no longer need to run on IBM Cell chip).
git-svn-id: http://svn.code.sf.net/p/petavision/code/trunk@2073 7c0444ca-5550-0410-9008-faecdbe78eb5

+++ src/layers/PVLayer.h
@@ -84,9 +84,8 @@ int pvlayer_finalize(PVLayer * l);
 
 int pvlayer_copyUpdate(PVLayer * l);
 
-// static, hopefully fast, routines:
 #ifdef DEPRECATED
-static inline int pvlayer_getPos(PVLayer * l, int k, float * x, float * y, float * kf)
+static inline int pvlayer_getPos(PVLayer * l, int k, float * x, float * y, int * kf)
 {
    *x = xPos(k, l->xOrigin, l->dx, l->loc.nx, l->loc.ny, l->numFeatures);
    *y = yPos(k, l->yOrigin, l->dy, l->loc.nx, l->loc.ny, l->numFeatures);
